Conan workspaces (#2933)
* some very basic conan-project core test

* working on it

* working...

* crazy local build

* more tests

* installer refactor

* one test failing

* tests pass

* installer.py refactor

* ongoing work

* working

* improving...

* super-project initial working

* minor cleanings

* lets share

* minor fix to default package folder

* changed behavior to build local as default

* added msg

* conditionals for OS

* tests pass linux

* tests in Windows

* using abs paths to build dirs

* Merged 1.4.0

* out of source conan-project

* fixed

* fixed linux

* fix test with Debug type

* generating conaninfo too

* renamed to workspace, using targets

* insource, multiple build systems, tests

* more timers

* even more timers

* larger timers

* super timers

* merged 1.4.4

* fix bad rename

* refactor installer project

* skip workspace test in OSX

+++ conans/client/installer.py
@@ -27,6 +27,7 @@ from conans.client.importer import remove_imports
 from conans.util.tracer import log_package_built
 from conans.client.tools.env import pythonpath
 from conans.client.package_installer import get_package
+from conans.client.file_copier import report_copied_files
 
 
 def build_id(conan_file):
@@ -200,7 +201,7 @@ def _handle_system_requirements(conan_file, package_reference, client_cache, out
 
     try:
         ret = str(ret or "")
-    except:
+    except Exception:
         out.warn("System requirements didn't return a string")
         ret = ""
     if getattr(conan_file, "global_system_requirements", None):
@@ -247,6 +248,7 @@ class ConanInstaller(object):
         self._build_mode = build_mode
         self._built_packages = set()  # To avoid re-building twice the same package reference
         self._recorder = recorder
+        self._workspace = None
 
     def install(self, deps_graph, profile_build_requires, keep_build=False, update=False):
         """ given a DepsGraph object, build necessary nodes or retrieve them
@@ -332,6 +334,36 @@ class ConanInstaller(object):
             self._call_package_info(conan_file, package_folder)
             clean_dirty(package_folder)
 
+    def _handle_node_workspace(self, node, workspace_package, inverse_levels, deps_graph):
+        conan_ref, conan_file = node.conan_ref, node.conanfile
+        output = ScopedOutput("Workspace %s" % conan_ref.name, self._out)
+        include_dirs = workspace_package.includedirs
+        lib_dirs = workspace_package.libdirs
+        self._call_package_info(conan_file, workspace_package.package_folder)
+        if include_dirs:
+            conan_file.cpp_info.includedirs = include_dirs
+        if lib_dirs:
+            conan_file.cpp_info.libdirs = lib_dirs
+            # Make sure the folders exists, otherwise they will be filtered out
+            lib_paths = [os.path.join(conan_file.cpp_info.rootpath, p)
+                         if not os.path.isabs(p) else p for p in lib_dirs]
+            for p in lib_paths:
+                mkdir(p)
+
+        self._propagate_info(node, inverse_levels, deps_graph)
+
+        build_folder = workspace_package.build_folder
+        write_generators(conan_file, build_folder, output)
+        save(os.path.join(build_folder, CONANINFO), conan_file.info.dumps())
+        output.info("Generated %s" % CONANINFO)
+        save(os.path.join(build_folder, BUILD_INFO), TXTGenerator(conan_file).content)
+        output.info("Generated %s" % BUILD_INFO)
+        # Build step might need DLLs, binaries as protoc to generate source files
+        # So execute imports() before build, storing the list of copied_files
+        from conans.client.importer import run_imports
+        copied_files = run_imports(conan_file, build_folder, output)
+        report_copied_files(copied_files, output)
+
     def _build(self, nodes_to_process, deps_graph, skip_nodes, profile_build_requires, keep_build,
                root_node, update):
         """ The build assumes an input of conans ordered by degree, first level
@@ -344,8 +376,13 @@ class ConanInstaller(object):
             level[:] = [n for n in level if n not in skip_nodes]
 
         for node, package_id, build_needed in nodes_to_process:
-            self._handle_node_cache(node, package_id, build_needed, keep_build, profile_build_requires,
-                                    inverse_levels, deps_graph, update)
+
+            workspace_package = self._workspace[node.conan_ref] if self._workspace else None
+            if workspace_package:
+                self._handle_node_workspace(node, workspace_package, inverse_levels, deps_graph)
+            else:
+                self._handle_node_cache(node, package_id, build_needed, keep_build, profile_build_requires,
+                                        inverse_levels, deps_graph, update)
 
         # Finally, propagate information to root node (conan_ref=None)
         self._propagate_info(root_node, inverse_levels, deps_graph)
@@ -486,18 +523,26 @@ class ConanInstaller(object):
                     package_folder = self._client_cache.package(package_reference,
                                                                 short_paths=conan_file.short_paths)
 
-                    with self._client_cache.package_lock(package_reference):
-                        if is_dirty(package_folder):
-                            output = ScopedOutput(str(conan_ref), self._out)
-                            output.warn("Package is corrupted, removing folder: %s" % package_folder)
-                            rmdir(package_folder)
+                    local_project = self._workspace[conan_ref] if self._workspace else None
+                    if not local_project:
+                        with self._client_cache.package_lock(package_reference):
+                            if is_dirty(package_folder):
+                                output = ScopedOutput(str(conan_ref), self._out)
+                                output.warn("Package is corrupted, removing folder: %s" % package_folder)
+                                rmdir(package_folder)
+
                     check_outdated = self._build_mode.outdated
                     if self._build_mode.forced(conan_file, conan_ref):
                         build_node = True
                     else:
-                        available = self._remote_proxy.package_available(package_reference, package_folder,
-                                                                         check_outdated)
-                        build_node = not available
+                        if local_project:
+                            # Maybe the folder is not there, check it!
+                            build_node = False
+                        else:
+                            available = self._remote_proxy.package_available(package_reference,
+                                                                             package_folder,
+                                                                             check_outdated)
+                            build_node = not available
 
                 nodes_to_build.append((node, package_id, build_node))
